Crate narwhal_storage

Structs

CertificateStore

The main storage when we have to deal with certificates. It maintains two storages, one main which saves the certificates by their ids, and a secondary one which acts as an index to allow us fast retrieval based for queries based in certificate rounds. It also offers pub/sub capabilities in write events. By using the notify_read someone can wait to hear until a certificate by a specific id has been written in storage.

CertificateStoreCache

An LRU cache for the certificate store.

CertificateStoreCacheMetrics

ConsensusStore

The persistent storage of the sequencer.

NodeStorage

All the data stores of the node.

PayloadStore

Store of the batch digests for the primary node for the own created batches.

ProposerStore

The storage for the proposer

VoteDigestStore

The storage for the last votes digests per authority

Constants

LAST_PROPOSAL_KEY

Traits

Cache

A cache trait to be used as temporary in-memory store when accessing the underlying certificate_store. Using the cache allows to skip rocksdb access giving us benefits both on less disk access (when value not in db’s cache) and also avoiding any additional deserialization costs.

Type Aliases

PayloadToken

ProposerKey

StoreResult

Convenience type to propagate store errors.