Trait narwhal_storage::Cache

pub trait Cache {
    // Required methods
    fn write(&self, certificate: Certificate);
    fn write_all(&self, certificate: Vec<Certificate>);
    fn read(&self, digest: &CertificateDigest) -> Option<Certificate>;
    fn read_all(
        &self,
        digests: Vec<CertificateDigest>
    ) -> Vec<(CertificateDigest, Option<Certificate>)>;
    fn contains(&self, digest: &CertificateDigest) -> bool;
    fn multi_contains<'a>(
        &self,
        digests: impl Iterator<Item = &'a CertificateDigest>
    ) -> Vec<bool>;
    fn remove(&self, digest: &CertificateDigest);
    fn remove_all(&self, digests: Vec<CertificateDigest>);
}

A cache trait to be used as temporary in-memory store when accessing the underlying certificate_store. Using the cache allows to skip rocksdb access giving us benefits both on less disk access (when value not in db’s cache) and also avoiding any additional deserialization costs.

Required Methods

fn write(&self, certificate: Certificate)

fn write_all(&self, certificate: Vec<Certificate>)

fn read(&self, digest: &CertificateDigest) -> Option<Certificate>

fn read_all( &self, digests: Vec<CertificateDigest> ) -> Vec<(CertificateDigest, Option<Certificate>)>

Returns the certificates by performing a look up in the cache. The method is expected to always return a result for every provided digest (when found will be Some, None otherwise) and in the same order.

fn contains(&self, digest: &CertificateDigest) -> bool

Checks existence of one or more digests.

fn multi_contains<'a>( &self, digests: impl Iterator<Item = &'a CertificateDigest> ) -> Vec<bool>

fn remove(&self, digest: &CertificateDigest)

fn remove_all(&self, digests: Vec<CertificateDigest>)

Object Safety

This trait is not object safe.